2019.03.19 21:48

+++ angular-src/src/app/components/board/board.component.ts
@@ -9,6 +9,8 @@ import { AuthService } from '../../services/auth.service';
 import { BoardService } from '../../services/board.service';
 import { DomSanitizer } from '@angular/platform-browser';
 import { NgFlashMessageService } from 'ng-flash-messages';
+import { Observable } from 'rxjs';
+import { FormControl } from '@angular/forms';
 
 @Component({
   selector: 'app-board',
@@ -18,12 +20,16 @@ import { NgFlashMessageService } from 'ng-flash-messages';
 @Pipe({ name: 'safeHtml' })
 export class BoardComponent implements OnInit, OnDestroy, PipeTransform {
   navigationSubscription;
+  filteredOptions: Observable<string[]>;
+
   type: String;
   num: String;
   content: any;
   contents: Object[];
-  cmtWrite: String;
+  cmtWrite = new FormControl();
   user: any;
+
+  ct: any = null;
 
   constructor(
     private route: ActivatedRoute,
@@ -76,6 +82,7 @@ export class BoardComponent implements OnInit, OnDestroy, PipeTransform {
     this.type = this.route.snapshot.paramMap.get('type');
     this.num = this.route.snapshot.paramMap.get('num');
 
+    this.setFilteredOptions();
     this.boardService.takeAllPosts(this.type).subscribe(data => {
       this.contents = data.posts;
       if ( this.num != 'list' ) {
@@ -103,35 +110,82 @@ export class BoardComponent implements OnInit, OnDestroy, PipeTransform {
     }
   }
 
+  private setFilteredOptions() {
+    this.filteredOptions = this.cmtWrite.valueChanges.pipe(
+      // startWith('TO::'),
+      map(value => this._filter(value))
+    );
+  }
+
+  private _filter(value: string): any[] {
+    const filterValue = value.toLowerCase();
+
+    // return this.options.filter(option => option.toLowerCase().includes(filterValue));
+
+    // let f = [];
+    // for (let i=0; i<this.foods.length; i++) {
+    //   f[i] = 'TO::' + this.foods[i].viewValue + ' [' + this.foods[i].value + '] ';
+    // }
+    // return f.filter(option => option.toLowerCase().includes(filterValue));
+    // return this.foods;
+
+    return this.content.comment;
+  }
+
+  setCmtTo(ct: any) {
+    this.ct = ct;
+  }
+
   onWriteComment() {
-    let headers = new Headers();
-    headers.append('Content-Type', 'application/json');
-    let formData = {
-      comment: this.cmtWrite
+    const formData = {
+      comment: this.cmtWrite.value,
+      _id: this.num
     };
-    this.http.post(this.funcService.ServerAddress + '/boards/writeComment', formData, {headers: headers}).pipe(map((res: Response) => res.json())).subscribe((data: ResponseWriting) => {
+    this.boardService.writeComment(formData).subscribe(data => {
+      console.log(data);
       if ( data.success ) {
+        this.router.navigate(['/board/' + this.type + '/' + this.num]);
         // i'd like to make this comment part to be working as an async way
+      } else {
+        this.flashMessage.showFlashMessage({
+          messages: ['댓글 작성 오류'], 
+          type: 'danger', 
+          timeout: 3000
+        });
       }
     });
   }
 
   onRemovePost() {
-    this.boardService.removePost(this.num).subscribe(result => {
-      if ( result.success ) {
-        this.router.navigate(['/board/' + this.type + '/list']);
-      } else {
-        this.flashMessage.showFlashMessage({
-          messages: ['삭제 오류'], 
-          type: 'danger', 
-          timeout: 3000
-        });
-      }
-    })
+    if ( confirm('삭제하시겠습니까?') ) {
+      this.boardService.removePost(this.num).subscribe(result => {
+        if ( result.success ) {
+          this.router.navigate(['/board/' + this.type + '/list']);
+        } else {
+          this.flashMessage.showFlashMessage({
+            messages: ['삭제 오류'], 
+            type: 'danger', 
+            timeout: 3000
+          });
+        }
+      })
+    }
   }
 
-}
+  onRemoveComment(cmtNum) {
+    if ( confirm('삭제하시겠습니까?') ) {
+      this.boardService.removeComment(this.num, cmtNum).subscribe(result => {
+        if ( result.success ) {
+          this.router.navigate(['/board/' + this.type + '/' + this.num]);
+        } else {
+          this.flashMessage.showFlashMessage({
+            messages: ['삭제 오류'], 
+            type: 'danger', 
+            timeout: 3000
+          });
+        }
+      })
+    }
+  }
 
-interface ResponseWriting {
-  success: Boolean;
 }
